// ui/src/generated/GrooveCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Beats } from "./Beats";
import type { ClipId } from "./ClipId";
import type { NoteId } from "./NoteId";

export type GrooveCommand = { "type": "Humanize", clip: ClipId, notes: Array<NoteId> | null, timing: Beats, velocity: number, seed: number, } | { "type": "SetSwing", amount: number, grid: Beats, };